Administrative Courts + Organic Law
Organic Law is a layer of legislation between Constitutional and Statutory law. So it goes:

1. Constitutional Law (Primary Law)
2. Organic Law (Secondary Law)
3. Statutory Law (Tertiary Law)
4. Delegatory Law (Quaternary Law)

Administrative Courts are essential here. Administrative Law is currently stifled by quite a lot of political meddling due to being shoved in the executive. This will allow it to flourish independently and restore separation of powers.